Fix homework 6

+++ 6_python/python_task3/python_task3.py
@@ -1,4 +1,4 @@
-tempC = int(input("Введите температуру в градусах Цельсия: "))
+tempC = float(input("Введите температуру в градусах Цельсия: "))
 tempF = float((tempC * 9/5) + 32)
 
 print (f"{tempC} °C = {tempF} °F")
